+++ src/070-optimizer/algebraic.ts
