+++ src/070-optimizer/associative.ts
